✨ Added a board representation
Instead of being 81 optional digits, this board representation
is stored as 9 bitfields, with 81 bits in each. This makes it
easier to represent invalid data (there can be multiple digits
in a single cell, and this needs to be accounted for) but it
makes it faster to do all of the calculations.